requirements.txt
Pillow
opencv-python-headless
certifi
boto3

numpy

pytesseract
